Updated index.html and layout.html
Added materialbox for the marker images so that they are brought to the front whenever you click them. Fixed the issue of the materialbox breaking the scrollbar on the edit mode for markers. Added a ribbon at the top to display "Marker information" instead of having that in the actual marker information. Re-did the X button and gave it a fixed place at the top of the marker popup.  Added the javascript to remove the button from view whenever you click the materialbox image, had to do this since the button was in a fixed place.