Updated index.html, layout.html
Index.html:
Removed the whitespace between the image and the ribbon in building info
Added javascript code for a hide and unhide button, added a hide and unhide button to move the respective pop-ups out of the way so that when someone is editing, the pop-up does not block the entire map. This was implemented mainly for phone.

Layout.html:
Removed the nav-bar as it wasn't necessary, changed some fonts, hid the original login details as it was blocking view and functions on phone, swapped the get location button with the layer toggle button as it was not moving since it was fixed.